refactor and separate GUI components related to widgets based on resolwe platform
